docs: update go docs
